Add space-spanning points to initial design
Sometimes the solution path from CRS2 concentrates around/along a boundary. Subsequent tools assume a certain level of diversity in the supplied design. Adding additional bad, space-spanning points should not worse optima estimation, but will slow down the GP portion of the optimiser,

An additional, and as yet unresolvable issue, is the sensitivity of CRS2 to the supplied initial value. We aren't using it for its intended use case (there is noise in the objective), but it would be nice to find some way to average over initial values / ameliorate this difference.